+++ dropwizard-guicey/src/test/groovy/ru/vyarus/dropwizard/guice/yaml/qualifier/QualifiedSampleNullValuesTest.groovy
@@ -3,7 +3,7 @@ package ru.vyarus.dropwizard.guice.yaml.qualifier
 import com.google.inject.Inject
 import com.google.inject.name.Named
 import io.dropwizard.metrics.common.MetricsFactory
-import jakarta.annotation.Nullable
+import javax.annotation.Nullable
 import ru.vyarus.dropwizard.guice.test.jupiter.TestGuiceyApp
 import spock.lang.Specification
 
@@ -20,9 +20,10 @@ class QualifiedSampleNullValuesTest extends Specification {
     @Inject @Named("sub-prop") Set<String> prop23
     @Inject @Named("metrics") MetricsFactory metricsFactyry
 
-    // both annotations work
+    // all tree annotations work
     @Inject @Nullable @jakarta.inject.Named("ee") String ee
-    @Inject @Nullable @Named("ee") String ee2
+    @Inject @Nullable @javax.inject.Named("ee2") String ee2
+    @Inject @Nullable @Named("ee") String ee3
 
     def "Check qualified bindings"() {
 
@@ -32,6 +33,7 @@ class QualifiedSampleNullValuesTest extends Specification {
         obj1.prop2 == null
         ee == null
         ee2 == null
+        ee3 == null
         prop23 == [null] as Set
         metricsFactyry != null
     }